+++ app/(root)/example/page.tsx
@@ -2,29 +2,20 @@
 import CButton from "@/components/shared/custome/c-button";
 import store from "@/store";
 
-import { useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
-import * as z from "zod";
-import { ProfileSchema } from "@/lib/validation";
-import CInput from "@/shared/custome/c-input";
-import { Form } from "@/components/ui/form";
-import { IoIosAdd } from "react-icons/io";
-import { Default } from "@/lib/image";
-import { Suspense, useState } from "react";
-import CInputV2 from "@/shared/custome/c-input-v2";
+import { Suspense } from "react";
 import DynamicList from "@/components/shared/dynamic-list";
 import DemoCarousel from "@/components/module/example/demo-carousel";
 import Pagination from "@/components/shared/pagination";
 import DemoSelect from "@/components/module/example/demo-select";
+import DemoInput from "@/components/module/example/demo-input";
 
 export default function Page() {
-  const { setOpenDrawer } = store();
-  const [search, setSearch] = useState("");
+  const { setOpenDialog } = store();
 
   const handleConfirmation = () => {
-    setOpenDrawer({
+    setOpenDialog({
       id: "CONFRIMATION",
-      // width: "w-[300px]",
+      width: "w-[300px]",
       height: "h-[200px]",
       headerTitle: "Confirmation",
       data: {
@@ -34,22 +25,8 @@ export default function Page() {
     });
   };
 
-  const form = useForm<z.infer<typeof ProfileSchema>>({
-    resolver: zodResolver(ProfileSchema),
-    defaultValues: {
-      name: "Jhon",
-    },
-  });
-
-  const nameValue = form.watch("name");
-  console.log(nameValue);
-
-  const onSubmit = async (values: z.infer<typeof ProfileSchema>) => {
-    console.log(values);
-  };
-
   return (
-    <div className="p-2 container mx-auto w-full">
+    <div className="p-2 container mx-auto w-full grid gap-10">
       <CButton
         title="Confirmation"
         onClick={handleConfirmation}
@@ -60,36 +37,7 @@ export default function Page() {
         item={["Item 1", "Item 2", "Item 3"]}
         render={(item) => <div key={item}>{item}</div>}
       />
-      <div className="flex items-start flex-col gap-4 max-w-xl">
-        <Form {...form}>
-          <form
-            onSubmit={form.handleSubmit(onSubmit)}
-            className="flex flex-col w-full"
-          >
-            <CInput
-              iconSvg={<IoIosAdd />}
-              label="Name"
-              placeholder="Jhon Doe"
-              form={form}
-              name="name"
-            />
-            <CInput
-              iconImg={Default}
-              label="Username"
-              placeholder="johndoe24"
-              form={form}
-              name="username"
-            />
-            <CButton title="Submit" className="mt-5" font="bd" size="sm" />
-          </form>
-        </Form>
-        <CInputV2
-          placeholder="Search..."
-          value={search}
-          name="search"
-          onChange={(e: any) => setSearch(e.target.value)}
-        />
-      </div>
+      <DemoInput />
       <DemoCarousel />
       <Suspense>
         <Pagination totalPages={20} />