+++ components/shared/custome/c-input.tsx
@@ -17,8 +17,9 @@ export default function CInput({
   iconSvg,
   iconImg,
   iconSize,
+  isForm,
 }: CInput) {
-  return (
+  return isForm ? (
     <FormField
       control={form.control}
       name={name}
@@ -27,40 +28,87 @@ export default function CInput({
         return (
           <FormItem>
             <FormControl>
-              <div className={cn("w-full", classNameParent)}>
-                {label && <Label>{label}</Label>}
-                <div className="relative">
-                  <div className="absolute top-1/2 left-3 -translate-y-1/2">
-                    {iconSvg && iconSvg}
-                    {iconImg && (
-                      <Image
-                        src={iconImg}
-                        alt="icon"
-                        width={iconSize || 20}
-                        height={iconSize || 20}
-                        className="flex-shrink-0"
-                      />
-                    )}
-                  </div>
-                  <Input
-                    placeholder={`${placeholder} ${required ? "*" : ""}`}
-                    className={cn(className, {
-                      "pl-7": iconSvg,
-                      "pl-10": iconImg,
-                    })}
-                    type={type}
-                    onWheel={(event) => event.currentTarget.blur()}
-                    checked={checked}
-                    isError={isError}
-                    {...field}
-                  />
-                </div>
-              </div>
+              <BaseInput
+                field={field}
+                isError={isError}
+                classNameParent={classNameParent}
+                className={className}
+                label={label}
+                iconSvg={iconSvg}
+                iconImg={iconImg}
+                iconSize={iconSize}
+                name={name}
+                placeholder={placeholder}
+                required={required}
+                type={type}
+                checked={checked}
+              />
             </FormControl>
             <FormMessage />
           </FormItem>
         );
       }}
     />
+  ) : (
+    <BaseInput
+      classNameParent={classNameParent}
+      className={className}
+      label={label}
+      iconSvg={iconSvg}
+      iconImg={iconImg}
+      iconSize={iconSize}
+      name={name}
+      placeholder={placeholder}
+      required={required}
+      type={type}
+      checked={checked}
+    />
   );
 }
+
+const BaseInput = ({
+  className,
+  classNameParent,
+  label,
+  iconSvg,
+  iconImg,
+  iconSize,
+  placeholder,
+  required,
+  type,
+  checked,
+  isError,
+  ...field
+}: CInput) => {
+  return (
+    <div className={cn("w-full", classNameParent)}>
+      {label && <Label>{label}</Label>}
+      <div className="relative">
+        <div className="absolute top-1/2 left-3 -translate-y-1/2">
+          {iconSvg && iconSvg}
+          {iconImg && (
+            <Image
+              src={iconImg}
+              alt="icon"
+              width={iconSize || 20}
+              height={iconSize || 20}
+              className="flex-shrink-0"
+            />
+          )}
+        </div>
+        <Input
+          placeholder={`${placeholder} ${required ? "*" : ""}`}
+          className={cn(className, {
+            "pl-7": iconSvg,
+            "pl-10": iconImg,
+          })}
+          type={type}
+          onWheel={(event) => event.currentTarget.blur()}
+          checked={checked}
+          isError={isError}
+          {...field}
+        />
+      </div>
+    </div>
+  );
+};