+++ CHANGELOG.md
@@ -63,3 +63,9 @@
 ### Added
 
 - custome select
+
+### Change
+
+- (Refactor) make custome input only 1
+- (Refactor) change color input, select, drawer and dialog
+- (Refactor) adding Esc and outside click to close drawer